Add time series

+++ fs/levels_manager_test.go
@@ -15,7 +15,7 @@ func TestMultiLevelFs(t *testing.T) {
 	cfg := db.NewDefaultConfig()
 	// fs := mockFilesystem[int32, *db.Kv]{}
 
-	levels, err := NewLeveledFilesystem[int32, *db.Kv](cfg)
+	levels, err := NewLeveledFilesystem[int64, *db.Kv](cfg)
 	assert.NoError(t, err)
 	_ = levels
 }
